fixed crashing

+++ src/info.h
@@ -1256,12 +1256,12 @@ typedef enum
 typedef struct
 {
   spritenum_t sprite;       // sprite number to show
-  long        frame;        // which frame/subframe of the sprite is shown
-  long        tics;         // number of gametics this frame should last
+  int         frame;        // which frame/subframe of the sprite is shown
+  int         tics;         // number of gametics this frame should last
   actionf_t   action;       // code pointer to function for action if any
   statenum_t  nextstate;    // linked list pointer to next state or zero
-  long        misc1, misc2; // used for psprite positioning
-  long        args[MAXSTATEARGS]; // [XA] mbf21 args
+  int         misc1, misc2; // used for psprite positioning
+  int         args[MAXSTATEARGS]; // [XA] mbf21 args
   byte       *tranmap; // ID24
   int         flags;
 } state_t;
